--- a/src/frontend/components/rd/Base/FeatureList/FeatureList.story.vue
+++ b/src/frontend/components/rd/Base/FeatureList/FeatureList.story.vue
@@ -1,6 +1,4 @@
-<script setup lang="ts">
-import { logEvent } from "histoire/client";
-</script>
+<script setup lang="ts"></script>
 
 <template>
   <Story title="Base/FeatureList" :layout="{ type: 'grid', width: '50%' }">