--- a/src/frontend/components/rd/Base/GoalCard/GoalCard.story.vue
+++ b/src/frontend/components/rd/Base/GoalCard/GoalCard.story.vue
@@ -1,6 +1,4 @@
-<script setup lang="ts">
-import { logEvent } from "histoire/client";
-</script>
+<script setup lang="ts"></script>
 
 <template>
   <Story title="Base/GoalCard" :layout="{ type: 'grid', width: '33%' }">