--- a/src/frontend/components/rd/Section/index.vue
+++ b/src/frontend/components/rd/Section/index.vue
@@ -28,7 +28,6 @@ withDefaults(defineProps<SectionProps>(), {
       :class="[
         {
           'space-y-4': spacing === 'sm',
-          'space-y-8': spacing === 'md',
           'space-y-8': spacing === 'lg', // Default
         },
         $props.class,